algorithm/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
pillow==10.1.0
dashscope==1.17.0
python-dotenv==1.0.0
reportlab==4.0.7
werkzeug==3.0.1